+++ .eslintrc.js
@@ -1,10 +1,10 @@
 module.exports = {
   root: true,
   env: {
-    browser: true
+    node: true
   },
   'extends': [
-    'plugin:vue/essential',
+    'plugin:vue/recommended',
     '@vue/standard'
   ],
   rules: {