+++ src/VuePlyr.vue
@@ -1,6 +1,6 @@
 <template>
   <div>
-    <slot/>
+    <slot />
   </div>
 </template>
 
@@ -8,7 +8,7 @@
   import Plyr from 'plyr'
 
   export default {
-    name: 'vue-plyr',
+    name: 'VuePlyr',
     props: {
       /** Options object for plyr config. **/
       options: {
@@ -30,9 +30,17 @@
         player: {}
       }
     },
+    computed: {
+      opts () {
+        const options = this.options
+        if (!this.options.hasOwnProperty('hideYouTubeDOMError')) {
+          options.hideYouTubeDOMError = true
+        }
+        return options
+      }
+    },
     mounted () {
       this.player = new Plyr(this.$el.firstChild, this.opts)
-      this.$emit('player', this.player)
       this.emit.forEach(element => {
         this.player.on(element, this.emitPlayerEvent)
       })
@@ -42,6 +50,7 @@
         this.player.destroy()
       } catch (e) {
         if (!(this.opts.hideYouTubeDOMError && e.message === 'The YouTube player is not attached to the DOM.')) {
+          // eslint-disable-next-line no-console
           console.error(e)
         }
       }
@@ -49,15 +58,6 @@
     methods: {
       emitPlayerEvent (event) {
         this.$emit(event.type, event)
-      }
-    },
-    computed: {
-      opts () {
-        const options = this.options
-        if (!this.options.hasOwnProperty('hideYouTubeDOMError')) {
-          options.hideYouTubeDOMError = true
-        }
-        return options
       }
     }
   }