+++ src/demo/nuxt/plugins/vue-plyr.js
@@ -1,4 +1,0 @@
-import Vue from 'vue'
-import VuePlyr from '../../../../dist/vue-plyr.es'
-
-Vue.use(VuePlyr)