+++ src/main.js
@@ -5,13 +5,11 @@ VuePlyr.install = (Vue, options = {}) => {
     VuePlyr.props.options.default = () => { return { ...options.plyr } }
   }
   if (options.emit) {
-    VuePlyr.props.emit.default = () => { return [ ...options.emit ] }
+    VuePlyr.props.emit.default = () => { return [...options.emit] }
   }
   Vue.component(VuePlyr.name, VuePlyr)
 }
 
-// Credit to https://github.com/irazasyed for this auto Vue.use() when
-// installing from unpkg or similar.
 if (typeof window !== 'undefined' && window.Vue) {
   window.Vue.use(VuePlyr)
 }