+++ tests/unit/presets/VimeoNP.vue
@@ -1,9 +1,0 @@
-<template>
-  <div data-plyr-provider="vimeo" data-plyr-embed-id="76979871"></div>
-</template>
-
-<script>
-  export default {
-    name: 'VimeoNP'
-  }
-</script>