+++ tests/unit/presets/YouTubeNP.vue
@@ -1,9 +1,0 @@
-<template>
-  <div data-plyr-provider="youtube" data-plyr-embed-id="bTqVqk7FSmY"></div>
-</template>
-
-<script>
-  export default {
-    name: 'YouTubeNP'
-  }
-</script>